(FM-7145) accelerate puppetserver_gem list
The puppetserver gem cli commands are particularly slow as they start a JVM.
Instead, for the list command, use the rubygems library from the puppet ruby,
setting the GEM_HOME and GEM_PATH to the values in the puppetserver conf file.

This commit also corrects various rubocop issues, and syncs applicable changes
made to the parent gem provider since this provider was copied from it.